requirements.txt
Django==3.2.*
uWSGI==2.0.*
celery==5.0.*
django-celery-results
mysqlclient==1.4.*
boto3==1.14.*
botocore==1.17.*
django-storages==1.11.*
djangorestframework==3.12.*
drf-spectacular